requirements.txt
notion-client==1.0.0
watchdog==3.0.0
markdown==3.4.3
python-frontmatter==1.0.0
click==8.1.3
pyyaml==6.0
pytest==7.3.1
python-dotenv==1.0.0